RTP: treat it as an encrypted protocol.
The goal is to have "RTP.ZOOM" flows classified as "Encrypted" and
not as "Cleartext".

To be fair, RTP might be in cleartext (meaning: you can play
the media stream without any other info than the flow itself) but nowadays
RTP is almost always encrypted (i.e SRTP), especially on the public
internet.

The other solution would be to add a special case for RTP in
`ndpi_is_encrypted_proto()` (plain RTP -> cleartext, RTP as master
-> encrypted, or depending on the app), but it seemed overwhelming.