Add generic functions for library configuration
This is a new way to provide (more) configuration options in nDPI.

The idea is to have a simple way to configure (most of) nDPI: only one
function (`ndpi_set_config()`) to set any configuration parameters
(in the present or on in the future) and this function prototype is as
agnostic as possible.

This way, anytime we need to add a new configuration parameter:
 * we don't need to add two public functions (a getter and a setter)
 * we don't break API/ABI compatibility of the library; even changing
the parameter type (from integer to a list of integer, for example)
doesn't break the compatibility.

Two examples of how to extend the configuration are provided:
 * the ability to enable/disable the extraction of the sha1 fingerprint
of the TLS certificates.
 * the upper limit on the number of packets per flow that will be subject
to inspection